Move new layout logic out into separate file
I shouldn't have needed to do this, but I branched the file of an old
commit by mistake :sheep:
